--- service/src/main/java/org/sunbird/util/user/UserConstants.java
@@ -1,9 +1,0 @@
-package org.sunbird.util.user;
-
-public final class UserConstants {
-
-  public static final String SIGNUP_TYPE = "signupType";
-  public static final String USER_TYPE = "userType";
-
-  private UserConstants() {}
-}
